install fish

+++ background.sh
@@ -5,10 +5,17 @@
 mkdir -p ~/.ssh && sudo rm -f ~/.ssh/known_hosts
 ssh-keyscan github.com > ~/.ssh/known_hosts
 
-sudo apt purge -y neovim* fish*
+if [[ -e "$HOME/.nix-profile/etc/profile.d/nix.sh" ]]; then
+    sudo apt purge -y neovim* fish*
 
-nix-env -iA nixpkgs.neovim nixpkgs.ripgrep nixpkgs.fd nixpkgs.fzf nixpkgs.fish
+    . "$HOME/.nix-profile/etc/profile.d/nix.sh"
 
-sudo chsh -s "$HOME/.nix-profile/bin/fish" "$USER"
+    nix-env -iA nixpkgs.neovim nixpkgs.ripgrep nixpkgs.fd nixpkgs.fzf nixpkgs.fish
 
-bash <(curl -s https://raw.githubusercontent.com/lunarvim/lunarvim/master/utils/installer/install.sh) -y
+    sudo chsh -s "$HOME/.nix-profile/bin/fish" "$USER"
+
+    fish -c 'curl -sL https://git.io/fisher | source && fisher install jorgebucaran/fisher'
+    fish -c 'fisher install PatrickF1/fzf.fish'
+
+    bash <(curl -s https://raw.githubusercontent.com/lunarvim/lunarvim/master/utils/installer/install.sh) -y
+fi
